Slice up snowflake.go
